requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
numpy>=1.24.0
pandas>=1.5.0
scikit-learn>=1.0.0
xgboost>=1.6.0
h5py>=3.8.0
python-dotenv>=0.19.0
requests>=2.26.0
joblib>=1.0.1 